+++ test/sample/FillBlanksPuzzleTest.java
@@ -70,7 +70,7 @@ class FillBlanksPuzzleTest {
     void checkFeedback() {
         List<Block> answers = new ArrayList<>(p.getSolutionSet());
 
-        int startTime = (int) System.currentTimeMillis();
+        long startTime = System.currentTimeMillis();
         assertEquals(p.getNumAttempts(), 0);
         assertFalse(p.isCompleted());
